wick-a11y voice support and violations summary for each test
