+++ .gitignore
@@ -9,11 +9,13 @@
 # Node.js dependencies
 /node_modules/
 
+# Doc videos
+/videos/
+
 # Environment variables
 /.env
 .env.local
 .env.*.local
-cypress.env.json
 
 # Editor-specific files
 /.vscode/
