+++ cypress.env.json
@@ -1,0 +1,3 @@
+{
+    "enableAccessibilityVoice": true
+}
