+++ cypress/e2e/all.js
@@ -5,8 +5,7 @@ import "../../src/accessibility-commands.js";
 
 
 describe('TEST SUITE 1', { tags: ['@accessibility'] }, () => {
-
-    // #6
+    
     it('Check accessibility Parabank web site - 6', { defaultCommandTimeout: 15000 }, () => {
         cy.visit('https://parabank.parasoft.com/parabank/index.htm')
         cy.wait(2000) // Using cy.wait(TIME) is a very bad practice, but it is used for simplicity in this example
@@ -15,7 +14,6 @@ describe('TEST SUITE 1', { tags: ['@accessibility'] }, () => {
     });
 
     context('Context 1', () => {
-        // #10
         it('Check accessibility Practice Software Testing web site - 10', { defaultCommandTimeout: 15000 }, () => {
             cy.visit('https://practicesoftwaretesting.com/#/')
             cy.wait(2000) // Using cy.wait(TIME) is a very bad practice, but it is used for simplicity in this example
@@ -23,7 +21,6 @@ describe('TEST SUITE 1', { tags: ['@accessibility'] }, () => {
             cy.checkAccessibility()
         });
 
-        // #11
         it('Simple failing test - 11', () => {
             cy.wrap(false).as('falseValue')
             cy.get('@falseValue').should('be.true')
@@ -32,7 +29,6 @@ describe('TEST SUITE 1', { tags: ['@accessibility'] }, () => {
     })
 
     context('Context 2', () => {
-        // #12
         it('Check accessibility Cypress.io web site - 12', { defaultCommandTimeout: 15000 }, () => {
             cy.visit('https://example.cypress.io/')
             cy.wait(2000) // Using cy.wait(TIME) is a very bad practice, but it is used for simplicity in this example
@@ -40,13 +36,11 @@ describe('TEST SUITE 1', { tags: ['@accessibility'] }, () => {
             cy.checkAccessibility(null, { includedImpacts: ['critical', 'serious', 'moderate', 'minor'] })
         });
 
-        // #13
         it.skip('Simple passing test - 13', () => {
             expect(true).to.be.true
         });
     })
 
-    // #7
     it('Check accessibility google.com web site - 7', { defaultCommandTimeout: 15000 }, () => {
         cy.visit('https://www.google.com')
         cy.wait(2000) // Using cy.wait(TIME) is a very bad practice, but it is used for simplicity in this example
@@ -54,12 +48,10 @@ describe('TEST SUITE 1', { tags: ['@accessibility'] }, () => {
         cy.checkAccessibility()
     });
 
-    // #8
     it('Another simple passing test - 8', () => {
         expect(true).to.be.true
     });
 
-    // #9
     it('Check accessibility Applitools web site - 9', { defaultCommandTimeout: 15000 }, () => {
         cy.visit('https://demo.applitools.com/')
         cy.wait(2000) // Using cy.wait(TIME) is a very bad practice, but it is used for simplicity in this example
@@ -72,22 +64,18 @@ describe('TEST SUITE 1', { tags: ['@accessibility'] }, () => {
 
 describe('TEST SUITE 2', { tags: ['@accessibility'] }, () => {
 
-    // #14
     it('Another simple passing test - 14', () => {
         expect(true).to.be.true
     });
 
-    // #15
     it('Another simple failing test - 15', () => {
         expect(true).to.be.false
     });
 
-    // #16
     it.skip('Another simple skipping test - 16', () => {
         expect(true).to.be.true
     });
 
-    // 17
     it('Check accessibility BrowserStack Demo web site - 17', { defaultCommandTimeout: 15000 }, () => {
         cy.visit('https://bstackdemo.com/')
         cy.wait(2000) // Using cy.wait(TIME) is a very bad practice, but it is used for simplicity in this example
@@ -97,7 +85,6 @@ describe('TEST SUITE 2', { tags: ['@accessibility'] }, () => {
 })
 
 
-// 1
 it('Check accessibility BrowserStack Demo web site - 1', { defaultCommandTimeout: 15000 }, () => {
     cy.visit('https://bstackdemo.com/')
     cy.wait(2000) // Using cy.wait(TIME) is a very bad practice, but it is used for simplicity in this example
@@ -105,18 +92,14 @@ it('Check accessibility BrowserStack Demo web site - 1', { defaultCommandTimeout
     cy.checkAccessibility(null, { includedImpacts: ['critical', 'serious', 'moderate', 'minor'] })
 });
 
-// 2
 it('Another simple passing test - 2', () => {
     expect(true).to.be.true
 });
 
-3
 it.skip('Another simple skipping test - 3', () => {
     expect(true).to.be.true
 });
 
-
-4
 it('Check accessibility BrowserStack Demo web site - 4', { defaultCommandTimeout: 15000 }, () => {
     cy.visit('https://bstackdemo.com/')
     cy.wait(2000) // Using cy.wait(TIME) is a very bad practice, but it is used for simplicity in this example
@@ -124,7 +107,6 @@ it('Check accessibility BrowserStack Demo web site - 4', { defaultCommandTimeout
     cy.checkAccessibility(null, { includedImpacts: ['critical', 'serious', 'moderate', 'minor'] })
 });
 
-// 5
 it('Another simple failing test - 5 (Ok in second attempt)', () => {
     if (Cypress.currentRetry === 1) {
         expect(true).to.be.true
